# docs/swagger.yaml
basePath: /v1/api
definitions:
...
  go-app_internal_dto.CreateChannelMemberDto:
    properties:
      channel_id:
        type: string
      role:
        type: string
      status:
        type: string
      user_ids:
        items:
          type: string
        type: array
    type: object
...
  go-app_internal_dto.UpdateChannelDto:
    properties:
      last_msg_id:
        type: string
      last_msg_seq:
        type: integer
      last_msg_time:
        type: string
    type: object
  go-app_internal_dto.UpdateChannelUnreadDto:
    properties:
      last_msg_id:
        type: string
      last_msg_time:
        type: string
      unread:
        type: integer
      version:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact: {}
  description: This is a sample server API for Go App.
  title: Go App API
  version: "1.0"
paths:
...
  /channel:
    get:
      description: Get a list of channels for the current user
      parameters:
      - description: Channel Type
        in: query
        name: type
        type: string
      - description: Updated At (RFC3339)
        in: query
        name: updated_at
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get channels
      tags:
      - channel
  /channel/{channel-id}:
    delete:
      description: Delete a channel by ID
      parameters:
      - description: Channel ID
        in: path
        name: channel-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete channel
      tags:
      - channel
    put:
      consumes:
      - application/json
      description: Update channel info
      parameters:
      - description: Channel ID
        in: path
        name: channel-id
        required: true
        type: string
      - description: Update Channel Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.UpdateChannelDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update channel
      tags:
      - channel
...
  /channel/{channel-id}/members:
    get:
      description: Get a list of members in a specific channel
      parameters:
      - description: Channel ID
        in: path
        name: channel-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get channel members
      tags:
      - channel
  /channel/{channel-id}/members/count:
    get:
      description: Get the number of members in a specific channel
      parameters:
      - description: Channel ID
        in: path
        name: channel-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get channel member count
      tags:
      - channel
...
  /channel/members:
    post:
      consumes:
      - application/json
      description: Add one or more members to a channel
      parameters:
      - description: Member Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.CreateChannelMemberDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Add member to channel
      tags:
      - channel
  /channel/members/{member-id}:
    delete:
      description: Remove a specific member from a channel
      parameters:
      - description: Member ID
        in: path
        name: member-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Remove member from channel
      tags:
      - channel
  /channel/unreads/{unread-id}:
    delete:
      description: Delete a channel unread record
      parameters:
      - description: Unread ID
        in: path
        name: unread-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete channel unread
      tags:
      - channel
    put:
      consumes:
      - application/json
      description: Update unread message count status
      parameters:
      - description: Unread ID
        in: path
        name: unread-id
        required: true
        type: string
      - description: Update Info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/go-app_internal_dto.UpdateChannelUnreadDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Update channel unread
      tags:
      - channel
  /channel/unreads/user/{user-id}:
    get:
      description: Get unread message counts for a user
      parameters:
      - description: User ID
        in: path
        name: user-id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get channel unreads
      tags:
      - channel
...
swagger: "2.0"